Move App.tsx into its own page
