+++ src/pages/App/App.scss
