+++ src/pages/App/App.tsx
@@ -1,8 +1,8 @@
-import logo from './assets/logo/logo.svg';
+import logo from '../../assets/logo/logo.svg';
 import './App.scss';
 import { BrowserRouter, Routes, Route, Link } from 'react-router-dom';
-import Swap from './pages/Swap';
-import Pool from './pages/Pool';
+import Swap from '../Swap';
+import Pool from '../Pool';
 
 function App() {
   return (
